requirements.txt
beautifulsoup4
html5lib
lxml
PyMuPDF
gspread==6.2.0
gspread_dataframe==4.0.0
oauth2client==4.1.3
openai==1.75.0
pandas==2.2.3
PyDrive2==1.21.3
python-dotenv==1.1.0
python-telegram-bot==22.0
pytz==2023.3.post1
seleniumbase
tiktoken==0.9.0
tqdm==4.65.0
